Squashed: Fixed the early release bug and cleaned up
Here's the issue: when using early release, we are depending on the
freelist itself for undoing the allocation.  Thus we have to be
careful not to overwrite the allocation information when freeing.
This will happen when the number of in-flight registers approach the
free registers and we have to reserve one more space than we need to
in the classic scheme.

Renamed the renamp table MAP rather than RAT.