First proof of concept of simplest first-principles OoO execution
Worst known issues:
- execution forwards in zero cycles to the next
- loads completely ignore pending stores and goes straight to memory

This branch is a reaction to the mess that was the previous spagetti
code; there were too much cached state and derived.  A real
implementation needs this, but here we try to go bottom up: what is
the simplest implementation of (inefficient) OoO?

We admit a structure (the ROB) to hold all the instructions we aim to
execute, some of which might have been fetch from a misspeculated
branch.  Instructions in the ROB enter in program order, youngest in,
and, eventually, oldest out.

We also need the architectural state of the register file
which is the complete state of registers when the ROB is empty.

Any instruction in the ROB may be executed if all its dependencies are
satisfied, typically, each source register either isn't written by an
older instruction in the ROB or if it is, then the youngest of those
have executed (COMMITTED).

For loads, we also need to check all potentially overlapping older
stores in the ROB. If we can tell none can overlap, the load can
proceed.  If all overlapping stores have committed, then we can
calculate the part that needs to be forwarded and combined load data
(if necessary).  Otherwise, we must wait.