[abstract exec] step 1 - 5
The idea here is that the decoded instruction should contain
everything to enable the execution without going back to the
instruction bits.  We want this as we may do macro fusion and other
fun things.  However this is large change so we are proceeding in
small steps (and this change is far from the end).